Add Lakka.svg
I had a go at making an svg of the logo so that we can render it to different resolutions.

The proportions might not be quite true to the original, I'm open to trying to fix issues or someone else tweaking it.

I've been told @kivutar may be the original author. Maybe there's already an svg somewhere that I couldn't find?